+++ Proyecto/src/assets/css/BancoPreguntas.css
@@ -1,20 +1,38 @@
 .app-container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
+  display: block;
+  align-items: flex-start;
+  margin: 40px auto 0;
+  padding: 0 20px;
+}
+
+.letra{
+  color: rgb(255, 255, 255);
+
+}
+.cont{
+  background-color: #5885ffc0;
+  border-radius: 8px;
+  border: 1px solid #000000;
+  padding: 30px;
+}
+
+.pregunta {
+  background: #002f80c4;
+  padding: 15px 35px;
+  margin-bottom: 15px;
+  border: 1px solid #000000;
 }
 
 .card {
-  background: #f9f9f9;
-  border-radius: 8px;
+  background: #002f8075;
   padding: 15px;
   margin-bottom: 15px;
-  border: 1px solid #ddd;
+  border: 1px solid #000000;
 }
 
 .card.correct {
-  border: 2px solid #4CAF50;
-  background: #f0fff0;
+  border: 10px solid #00be06be;
+  background: #002f8075;
 }
 
 label {
@@ -54,7 +72,7 @@ textarea, input[type="number"] {
 
 .select-button {
   padding: 8px 12px;
-  background: #e0e0e0;
+  background: #353535;
   border: none;
   border-radius: 4px;
   cursor: pointer;
@@ -69,22 +87,30 @@ textarea, input[type="number"] {
 .preview {
   margin-top: 30px;
   padding: 20px;
-  background: #f5f5f5;
+  background: #5885ffa9;
   border-radius: 8px;
 }
 
 .save-button {
   padding: 10px 20px;
-  background: #2196F3;
-  color: white;
+  background-color: #4caf50;
+  color: rgb(255, 255, 255);
   border: none;
   border-radius: 4px;
   cursor: pointer;
-  font-size: 16px;
-  margin-top: 15px;
+  font-size: 1rem;
+  width: 100%;
+  font-weight: 600;
 }
 
 .save-button:disabled {
-  background: #cccccc;
-  cursor: not-allowed;
+  padding: 10px 20px;
+  background-color: #4caf50;
+  color: rgb(255, 255, 255);
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 1rem;
+  width: 100%;
+  font-weight: 600;
 }
