+++ Proyecto/src/assets/css/CreadorPreguntas.css
@@ -1,16 +1,16 @@
 .app-container {
   max-width: 1200px;
-  margin: 0 auto;
+  margin: 40px auto 0;
   padding: 20px;
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 }
 
 h1, h2, h3 {
-  color: #2c3e50;
+  color: #141414;
 }
 
 .filters-section {
-  background-color: #f8f9fa;
+  background-color: #52b8f3cc;
   padding: 20px;
   border-radius: 8px;
   margin-bottom: 20px;
@@ -36,7 +36,7 @@ h1, h2, h3 {
 .filter-group input {
   width: 100%;
   padding: 8px;
-  border: 1px solid #ddd;
+  border: 1px solid #ffffff;
   border-radius: 4px;
 }
 
@@ -51,10 +51,10 @@ h1, h2, h3 {
 }
 
 .question-card {
-  border: 1px solid #e0e0e0;
+  border: 3px solid #000000ab;
   border-radius: 8px;
   padding: 15px;
-  background-color: white;
+  background-color: rgba(255, 255, 255, 0.842);
   box-shadow: 0 2px 4px rgba(0,0,0,0.05);
 }
 
@@ -84,17 +84,19 @@ h1, h2, h3 {
   list-style-type: none;
   padding: 0;
   margin: 10px 0;
+  color: #8f0000d7;
 }
 
 .options-list li {
   padding: 5px 10px;
   margin: 3px 0;
-  border-left: 3px solid #ddd;
+  border-left: 3px solid #8f0000d7;
 }
 
 .correct-option {
   border-left-color: #4caf50 !important;
   background-color: #f1f8e9;
+  color: #4caf50;
 }
 
 .question-footer {
@@ -129,7 +131,7 @@ h1, h2, h3 {
 }
 
 .test-section {
-  background-color: #f5f5f5;
+  background-color: #52b8f3cc;
   padding: 20px;
   border-radius: 8px;
 }
@@ -154,7 +156,7 @@ h1, h2, h3 {
 .form-group textarea {
   width: 100%;
   padding: 8px;
-  border: 1px solid #ddd;
+  border: 1px solid #000000;
   border-radius: 4px;
 }
 
@@ -165,13 +167,14 @@ h1, h2, h3 {
 
 .selected-questions {
   grid-column: span 2;
+  color: #000000;
 }
 
 .questions-list {
   max-height: 300px;
   overflow-y: auto;
-  background-color: white;
-  border: 1px solid #ddd;
+  background-color: rgba(255, 255, 255, 0.651);
+  border: 1px solid #000000;
   border-radius: 4px;
   padding: 10px;
   margin: 10px 0;
@@ -212,7 +215,7 @@ h1, h2, h3 {
 }
 
 .empty-message {
-  color: #757575;
+  color: #535353;
   text-align: center;
   padding: 20px;
 }
@@ -229,12 +232,13 @@ h1, h2, h3 {
 .save-test-btn {
   padding: 10px 20px;
   background-color: #4caf50;
-  color: white;
+  color: rgb(255, 255, 255);
   border: none;
   border-radius: 4px;
   cursor: pointer;
   font-size: 1rem;
   width: 100%;
+  font-weight: 600;
 }
 
 .save-test-btn:disabled {
