+++ Proyecto/src/assets/css/index.css
@@ -1,0 +1,93 @@
+:root {
+  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+  font-weight: 500;
+  color-scheme: light dark;
+  color: #213547;
+  background-color: #ffffff;
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+body {
+  margin: 0;
+  font-family: inherit;
+  min-width: 320px;
+  min-height: 100vh;
+  background-color: #ffffff;
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  background: url('./assets/images.jpeg') no-repeat center center fixed;
+  background-size: cover;
+  background-position: right -35px;
+}
+
+/* ===== NAVBAR PRINCIPAL ===== */
+.navbar {
+  background-color: rgba(255, 255, 255, 0.74); /* Blanco 85% opaco */
+  border-radius: 8px;
+  margin: auto;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.bottom-nav {
+  background-color: #002e8a;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 8px 40px;
+  font-weight: 600;
+}
+
+/* ===== NAV ITEMS ===== */
+.nav-link {
+  color: white;
+  background: none;
+  border: none;
+  margin-left: 20px;
+  font-size: 14px;
+  cursor: pointer;
+  text-decoration: none;
+  font-weight: 500;
+}
+
+.nav-link:hover {
+  text-decoration: underline;
+}
+
+.nav-link.active {
+  font-weight: 700;
+  text-decoration: underline;
+}
+
+/* ===== LOGO ===== */
+.logo {
+  display: flex;
+  align-items: center;
+}
+
+.logo img {
+  height: 30px;
+  margin-right: 10px;
+}
+
+/* ===== MODO OSCURO (opcional) ===== */
+@media (prefers-color-scheme: dark) {
+  :root {
+    color: #ffffff;
+    background-color: #1a1a1a;
+  }
+
+  .bottom-nav {
+    background-color: #003080;
+  }
+
+  .nav-link {
+    color: #ffffff;
+  }
+
+  .nav-link:hover {
+    color: #aaccff;
+  }
+}
