+++ Proyecto/src/index.jsx
@@ -1,11 +1,10 @@
-
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import Home from './Home.jsx'
+import './assets/css/index.css' //estilo global
+import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <Home />
+    <App />
   </StrictMode>
 )
