+++ Proyecto/src/pages/Home.jsx
@@ -1,0 +1,3 @@
+export default function Home() {
+  return null
+}
